Taking another approach in order to avoid cloning each rule after parsing    - parse all rules and store them in a HashMap    - save the hash to a file    - disambiguate
